% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_txi.R
\docType{data}
\name{syn_txi}
\alias{syn_txi}
\title{A dataset containing a txi object for 81 synovial biopsies from 
the PEAC cohort}
\format{
A txi object containing the count data for synovial samples
}
\source{
\url{https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1}
}
\usage{
syn_txi
}
\description{
A dataset containing a txi object for 81 synovial biopsies from 
the PEAC cohort
}
\keyword{datasets}
